% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.r
\name{hello_world}
\alias{hello_world}
\title{Hello World Function}
\usage{
hello_world(name = "World")
}
\arguments{
\item{name}{A word.}
}
\value{
A string.
}
\description{
Hello World Function
}
\examples{
hello_world()
hello_world('Tom')
}
